requirements.txt
streamlit
pandas
requests
beautifulsoup4
scikit-learn
pydantic
python-dotenv
openmeteo-requests
requests-cache
retry_requests
pytesseract
Pillow
